requirements.txt
flask==3.0.0
flask-login==0.6.3
python-dotenv==1.0.0
WooCommerce==3.0.0 